agents/programming/macOS Tahoe 26.2/requirements.txt
langchain
python-dotenv
requests
openai
# optional: if you have a local ollama integration package, add it here
# ollama-client
